fix(net)!: rename packet.rs to headers.rs
BREAKING_CHANGE: the net packet module is now named headers.

This is more accurate to the core function of the file.

It also opens up the module name `packet` to serve the actual function of a packet (rather than just header manipulation).
