+++ net/src/ipv6/mod.rs
@@ -3,12 +3,12 @@
 
 //! Ipv6 Address type and manipulation
 
+use crate::headers::Header;
 use crate::icmp6::Icmp6;
 use crate::ip::NextHeader;
 use crate::ip_auth::IpAuth;
 use crate::ipv6::addr::UnicastIpv6Addr;
 use crate::ipv6::flow_label::FlowLabel;
-use crate::packet::Header;
 use crate::parse::{
     DeParse, DeParseError, IntoNonZeroUSize, LengthError, Parse, ParseError, ParsePayload,
     ParsePayloadWith, ParseWith, Reader,
