+++ net/src/lib.rs
@@ -18,13 +18,13 @@ extern crate alloc;
 extern crate core;
 
 pub mod eth;
+pub mod headers;
 pub mod icmp4;
 pub mod icmp6;
 pub mod ip;
 pub mod ip_auth;
 pub mod ipv4;
 pub mod ipv6;
-pub mod packet;
 pub mod parse;
 pub mod tcp;
 pub mod udp;
